+++ Modules/Intent.Modules.VisualStudio.Projects/FactoryExtensions/NuGet/HelperTypes/NuGetProject.cs
@@ -16,7 +16,7 @@ internal class NuGetProject
     public string Content { get; set; }
     public Dictionary<string, NuGetPackage> RequestedPackages { get; set; } = new(StringComparer.OrdinalIgnoreCase);
     public Dictionary<string, NuGetPackage> InstalledPackages { get; set; } = new(StringComparer.OrdinalIgnoreCase);
-    public Dictionary<string, VersionRange> HighestVersions { get; set; } = new(StringComparer.OrdinalIgnoreCase);
+    public Dictionary<string, VersionInfo> HighestVersions { get; set; } = new(StringComparer.OrdinalIgnoreCase);
     public string FilePath { get; set; }
     public INuGetSchemeProcessor Processor { get; set; }
     public IOutputTarget OutputTarget { get; set; }
@@ -28,6 +28,6 @@ internal class NuGetProject
 
         return allPackages
             .GroupBy(x => x.Key, x => x.Value)
-            .ToDictionary(x => x.Key, x => x.OrderByDescending(y => y.Version?.MinVersion).First());
+            .ToDictionary(x => x.Key, x => x.OrderByDescending(y => y.Version).First());
     }
 }