+++ Modules/Intent.Modules.VisualStudio.Projects/FactoryExtensions/NuGet/PackageVersionRange.cs
@@ -2,4 +2,4 @@
 using NuGet.Versioning;
 
 namespace Intent.Modules.VisualStudio.Projects.FactoryExtensions.NuGet;
-internal record PackageVersionRange(VersionRange VersionRange, NuGetPackage? Package);
+internal record PackageVersionInfo(VersionInfo VersionInfo, NuGetPackage? Package);