+++ Modules/Intent.Modules.VisualStudio.Projects/release-notes.md
@@ -2,6 +2,7 @@
 
 - Fixed: Fixed Solution Relative Location path using `Path.GetFullPath` to correctly handle relative locations like ../.
 - Improvement: Updated `.gitignore` to exclude Intent Architect `*.application.log` files.
+- Improvement: Added support for variable syntax in package versions..
 
 ### Version 3.8.13